refactor: Separate graph custom function
